re-exported oox::node, refactored examples out to separate files, extended examples and benches
